working cli
