+++ cli/chex/noaka_symbols.txt
