+++ cli/chex/requirements.txt
@@ -1,0 +1,4 @@
+click==8.1.7
+prettytable==3.12.0
+setuptools==75.3.0
+lxml==5.1.0
